--- a/benchmarks/src/main/java/io/rsocket/PayloadsMaxPerfSubscriber.java
+++ b/benchmarks/src/main/java/io/rsocket/PayloadsMaxPerfSubscriber.java
@@ -1,10 +1,6 @@
 package io.rsocket;
 
 import org.openjdk.jmh.infra.Blackhole;
-import org.reactivestreams.Subscription;
-import reactor.core.CoreSubscriber;
-
-import java.util.concurrent.CountDownLatch;
 
 public class PayloadsMaxPerfSubscriber extends MaxPerfSubscriber<Payload> {