--- a/benchmarks/src/main/java/io/rsocket/internal/UnicastVsDefaultMonoProcessorPerf.java
+++ b/benchmarks/src/main/java/io/rsocket/internal/UnicastVsDefaultMonoProcessorPerf.java
@@ -1,7 +1,7 @@
 package io.rsocket.internal;
 
-
 import io.rsocket.MaxPerfSubscriber;
+import java.util.concurrent.TimeUnit;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
@@ -14,8 +14,6 @@ import org.openjdk.jmh.annotations.Warmup;
 import org.openjdk.jmh.infra.Blackhole;
 import reactor.core.publisher.MonoProcessor;
 
-import java.util.concurrent.TimeUnit;
-
 @BenchmarkMode({Mode.Throughput, Mode.SampleTime})
 @Fork(1)
 @Warmup(iterations = 10)
@@ -24,25 +22,25 @@ import java.util.concurrent.TimeUnit;
 @OutputTimeUnit(TimeUnit.MICROSECONDS)
 public class UnicastVsDefaultMonoProcessorPerf {
 
-    @Benchmark
-    public void monoProcessorPerf(Blackhole bh) {
-        MaxPerfSubscriber<Integer> subscriber = new MaxPerfSubscriber<>(bh);
-        MonoProcessor<Integer> monoProcessor = MonoProcessor.create();
-        monoProcessor.onNext(1);
-        monoProcessor.subscribe(subscriber);
+  @Benchmark
+  public void monoProcessorPerf(Blackhole bh) {
+    MaxPerfSubscriber<Integer> subscriber = new MaxPerfSubscriber<>(bh);
+    MonoProcessor<Integer> monoProcessor = MonoProcessor.create();
+    monoProcessor.onNext(1);
+    monoProcessor.subscribe(subscriber);
 
-        bh.consume(monoProcessor);
-        bh.consume(subscriber);
-    }
+    bh.consume(monoProcessor);
+    bh.consume(subscriber);
+  }
 
-    @Benchmark
-    public void unicastMonoProcessorPerf(Blackhole bh) {
-        MaxPerfSubscriber<Integer> subscriber = new MaxPerfSubscriber<>(bh);
-        UnicastMonoProcessor<Integer> monoProcessor = UnicastMonoProcessor.create();
-        monoProcessor.onNext(1);
-        monoProcessor.subscribe(subscriber);
+  @Benchmark
+  public void unicastMonoProcessorPerf(Blackhole bh) {
+    MaxPerfSubscriber<Integer> subscriber = new MaxPerfSubscriber<>(bh);
+    UnicastMonoProcessor<Integer> monoProcessor = UnicastMonoProcessor.create();
+    monoProcessor.onNext(1);
+    monoProcessor.subscribe(subscriber);
 
-        bh.consume(monoProcessor);
-        bh.consume(subscriber);
-    }
+    bh.consume(monoProcessor);
+    bh.consume(subscriber);
+  }
 }