--- a/rsocket-core/src/main/java/io/rsocket/internal/UnicastMonoProcessor.java
+++ b/rsocket-core/src/main/java/io/rsocket/internal/UnicastMonoProcessor.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2015-2019 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package io.rsocket.internal;
 
 import java.util.Objects;
@@ -10,7 +26,6 @@ import org.reactivestreams.Subscription;
 import reactor.core.CoreSubscriber;
 import reactor.core.Disposable;
 import reactor.core.Exceptions;
-import reactor.core.Fuseable;
 import reactor.core.Scannable;
 import reactor.core.publisher.Mono;
 import reactor.core.publisher.Operators;
@@ -19,7 +34,7 @@ import reactor.util.annotation.Nullable;
 import reactor.util.context.Context;
 
 public class UnicastMonoProcessor<O> extends Mono<O>
-    implements Processor<O, O>, CoreSubscriber<O>, Fuseable, Disposable, Fuseable.QueueSubscription<O>, Scannable {
+    implements Processor<O, O>, CoreSubscriber<O>, Disposable, Subscription, Scannable {
 
   /**
    * Create a {@link UnicastMonoProcessor} that will eagerly request 1 on {@link
@@ -32,74 +47,46 @@ public class UnicastMonoProcessor<O> extends Mono<O>
     return new UnicastMonoProcessor<>();
   }
 
-  /**
-   * Indicates this Subscription has no value and not requested yet.
-   */
-  static final int NO_SUBSCRIBER_NO_RESULT  = 0;
-  /**
-   * Indicates this Subscription has no value and not requested yet.
-   */
+  /** Indicates this Subscription has no value and not requested yet. */
+  static final int NO_SUBSCRIBER_NO_RESULT = 0;
+  /** Indicates this Subscription has no value and not requested yet. */
   static final int NO_SUBSCRIBER_HAS_RESULT = 1;
-  /**
-   * Indicates this Subscription has no value and not requested yet.
-   */
-  static final int NO_REQUEST_NO_RESULT     = 4;
-  /**
-   * Indicates this Subscription has a value but not requested yet.
-   */
-  static final int NO_REQUEST_HAS_RESULT    = 5;
-  /**
-   * Indicates this Subscription has been requested but there is no value yet.
-   */
-  static final int HAS_REQUEST_NO_RESULT    = 6;
-  /**
-   * Indicates this Subscription has both request and value.
-   */
-  static final int HAS_REQUEST_HAS_RESULT   = 7;
-  /**
-   * Indicates the Subscription has been cancelled.
-   */
-  static final int CANCELLED                = 8;
-  /**
-   * Indicates this Subscription is in fusion mode and is currently empty.
-   */
-  static final int FUSED_EMPTY              = 16;
-  /**
-   * Indicates this Subscription is in fusion mode and has an undelivered notification about present result.
-   */
-  static final int FUSED_HAS_RESULT         = 32;
-  /**
-   * Indicates this Subscription is in fusion mode and has a result.
-   */
-  static final int FUSED_READY              = 64;
-  /**
-   * Indicates this Subscription is in fusion mode and its value has been consumed.
-   */
-  static final int FUSED_CONSUMED           = 128;
+  /** Indicates this Subscription has no value and not requested yet. */
+  static final int NO_REQUEST_NO_RESULT = 4;
+  /** Indicates this Subscription has a value but not requested yet. */
+  static final int NO_REQUEST_HAS_RESULT = 5;
+  /** Indicates this Subscription has been requested but there is no value yet. */
+  static final int HAS_REQUEST_NO_RESULT = 6;
+  /** Indicates this Subscription has both request and value. */
+  static final int HAS_REQUEST_HAS_RESULT = 7;
+  /** Indicates the Subscription has been cancelled. */
+  static final int CANCELLED = 8;
 
   volatile int state;
+
   @SuppressWarnings("rawtypes")
   static final AtomicIntegerFieldUpdater<UnicastMonoProcessor> STATE =
       AtomicIntegerFieldUpdater.newUpdater(UnicastMonoProcessor.class, "state");
 
   volatile int once;
+
   @SuppressWarnings("rawtypes")
   static final AtomicIntegerFieldUpdater<UnicastMonoProcessor> ONCE =
-          AtomicIntegerFieldUpdater.newUpdater(UnicastMonoProcessor.class, "once");
+      AtomicIntegerFieldUpdater.newUpdater(UnicastMonoProcessor.class, "once");
 
   volatile Subscription subscription;
+
   @SuppressWarnings("rawtypes")
   static final AtomicReferenceFieldUpdater<UnicastMonoProcessor, Subscription> UPSTREAM =
-          AtomicReferenceFieldUpdater.newUpdater(
-                  UnicastMonoProcessor.class, Subscription.class, "subscription");
+      AtomicReferenceFieldUpdater.newUpdater(
+          UnicastMonoProcessor.class, Subscription.class, "subscription");
 
   CoreSubscriber<? super O> actual;
 
   Throwable error;
   O value;
 
-  UnicastMonoProcessor() {
-  }
+  UnicastMonoProcessor() {}
 
   @Override
   @NonNull
@@ -156,28 +143,16 @@ public class UnicastMonoProcessor<O> extends Mono<O>
   }
 
   /**
-   * Tries to emit the value and complete the underlying subscriber or
-   * stores the value away until there is a request for it.
-   * <p>
-   * Make sure this method is called at most once
+   * Tries to emit the value and complete the underlying subscriber or stores the value away until
+   * there is a request for it.
+   *
+   * <p>Make sure this method is called at most once
+   *
    * @param v the value to emit
    */
   private void complete(O v) {
-    for (;;) {
+    for (; ; ) {
       int state = this.state;
-
-      if (state == FUSED_EMPTY) {
-        setValue(v);
-        //sync memory since setValue is non volatile
-        if (STATE.compareAndSet(this, FUSED_EMPTY, FUSED_READY)) {
-          Subscriber<? super O> a = actual;
-          a.onNext(v);
-          a.onComplete();
-          return;
-        }
-        //refresh state if race occurred so we test if cancelled in the next comparison
-        state = this.state;
-      }
 
       // if state is >= HAS_CANCELLED or bit zero is set (*_HAS_VALUE) case, return
       if ((state & ~HAS_REQUEST_NO_RESULT) != 0) {
@@ -186,18 +161,22 @@ public class UnicastMonoProcessor<O> extends Mono<O>
         return;
       }
 
-      if (state == HAS_REQUEST_NO_RESULT && STATE.compareAndSet(this, HAS_REQUEST_NO_RESULT, HAS_REQUEST_HAS_RESULT)) {
-        this.value = null;
-        Subscriber<? super O> a = actual;
-        a.onNext(v);
-        a.onComplete();
-        return;
+      if (state == HAS_REQUEST_NO_RESULT) {
+        final Subscriber<? super O> a = actual;
+        if (STATE.compareAndSet(this, HAS_REQUEST_NO_RESULT, HAS_REQUEST_HAS_RESULT)) {
+          this.value = null;
+          a.onNext(v);
+          a.onComplete();
+          return;
+        }
       }
       setValue(v);
-      if (state == NO_REQUEST_NO_RESULT && STATE.compareAndSet(this, NO_REQUEST_NO_RESULT, NO_REQUEST_HAS_RESULT)) {
+      if (state == NO_REQUEST_NO_RESULT
+          && STATE.compareAndSet(this, NO_REQUEST_NO_RESULT, NO_REQUEST_HAS_RESULT)) {
         return;
       }
-      if (state == NO_SUBSCRIBER_NO_RESULT && STATE.compareAndSet(this, NO_SUBSCRIBER_NO_RESULT, NO_SUBSCRIBER_HAS_RESULT)) {
+      if (state == NO_SUBSCRIBER_NO_RESULT
+          && STATE.compareAndSet(this, NO_SUBSCRIBER_NO_RESULT, NO_SUBSCRIBER_HAS_RESULT)) {
         return;
       }
     }
@@ -205,61 +184,43 @@ public class UnicastMonoProcessor<O> extends Mono<O>
 
   /**
    * Tries to emit completion the underlying subscriber
-   * <p>
-   * Make sure this method is called at most once
+   *
+   * <p>Make sure this method is called at most once
    */
   private void complete() {
-    for (;;) {
+    for (; ; ) {
       int state = this.state;
-
-      if (state == FUSED_EMPTY) {
-        if (STATE.compareAndSet(this, FUSED_EMPTY, FUSED_CONSUMED)) {
-          Subscriber<? super O> a = actual;
-          a.onComplete();
-          return;
-        }
-        //refresh state if race occurred so we test if cancelled in the next comparison
-        state = this.state;
-      }
 
       // if state is >= HAS_CANCELLED or bit zero is set (*_HAS_VALUE) case, return
       if ((state & ~HAS_REQUEST_NO_RESULT) != 0) {
         return;
       }
 
-      if ((state == HAS_REQUEST_NO_RESULT || state == NO_REQUEST_NO_RESULT) && STATE.compareAndSet(this, state, HAS_REQUEST_HAS_RESULT)) {
-        Subscriber<? super O> a = actual;
-        a.onComplete();
-        return;
+      if (state == HAS_REQUEST_NO_RESULT || state == NO_REQUEST_NO_RESULT) {
+        final Subscriber<? super O> a = actual;
+        if (STATE.compareAndSet(this, state, HAS_REQUEST_HAS_RESULT)) {
+          a.onComplete();
+          return;
+        }
       }
-      if (state == NO_SUBSCRIBER_NO_RESULT && STATE.compareAndSet(this, NO_SUBSCRIBER_NO_RESULT, NO_SUBSCRIBER_HAS_RESULT)) {
+      if (state == NO_SUBSCRIBER_NO_RESULT
+          && STATE.compareAndSet(this, NO_SUBSCRIBER_NO_RESULT, NO_SUBSCRIBER_HAS_RESULT)) {
         return;
       }
     }
   }
 
   /**
-   * Tries to emit error the underlying subscriber or
-   * stores the value away until there is a request for it.
-   * <p>
-   * Make sure this method is called at most once
+   * Tries to emit error the underlying subscriber or stores the value away until there is a request
+   * for it.
+   *
+   * <p>Make sure this method is called at most once
+   *
    * @param e the error to emit
    */
   private void complete(Throwable e) {
-    for (;;) {
+    for (; ; ) {
       int state = this.state;
-
-      if (state == FUSED_EMPTY) {
-        setError(e);
-        //sync memory since setValue is non volatile
-        if (STATE.compareAndSet(this, FUSED_EMPTY, FUSED_CONSUMED)) {
-          Subscriber<? super O> a = actual;
-          a.onError(e);
-          return;
-        }
-        //refresh state if race occurred so we test if cancelled in the next comparison
-        state = this.state;
-      }
 
       // if state is >= HAS_CANCELLED or bit zero is set (*_HAS_VALUE) case, return
       if ((state & ~HAS_REQUEST_NO_RESULT) != 0) {
@@ -267,12 +228,15 @@ public class UnicastMonoProcessor<O> extends Mono<O>
       }
 
       setError(e);
-      if ((state == HAS_REQUEST_NO_RESULT || state == NO_REQUEST_NO_RESULT) && STATE.compareAndSet(this, state, HAS_REQUEST_HAS_RESULT)) {
-        Subscriber<? super O> a = actual;
-        a.onError(e);
-        return;
+      if (state == HAS_REQUEST_NO_RESULT || state == NO_REQUEST_NO_RESULT) {
+        final Subscriber<? super O> a = actual;
+        if (STATE.compareAndSet(this, state, HAS_REQUEST_HAS_RESULT)) {
+          a.onError(e);
+          return;
+        }
       }
-      if (state == NO_SUBSCRIBER_NO_RESULT && STATE.compareAndSet(this, NO_SUBSCRIBER_NO_RESULT, NO_SUBSCRIBER_HAS_RESULT)) {
+      if (state == NO_SUBSCRIBER_NO_RESULT
+          && STATE.compareAndSet(this, NO_SUBSCRIBER_NO_RESULT, NO_SUBSCRIBER_HAS_RESULT)) {
         return;
       }
     }
@@ -285,11 +249,13 @@ public class UnicastMonoProcessor<O> extends Mono<O>
     if (once == 0 && ONCE.compareAndSet(this, 0, 1)) {
       this.actual = actual;
 
-      // CAS LOOP since there is a racing with [onNext / onComplete / onError / dispose] which can appear at any moment
+      // CAS LOOP since there is a racing with [onNext / onComplete / onError / dispose] which can
+      // appear at any moment
 
       int state = this.state;
 
-      // possible states within the racing between [onNext / onComplete / onError / dispose] and setting subscriber
+      // possible states within the racing between [onNext / onComplete / onError / dispose] and
+      // setting subscriber
       // are NO_SUBSCRIBER_[NO_RESULT or HAS_RESULT]
       if (state == NO_SUBSCRIBER_NO_RESULT) {
         if (STATE.compareAndSet(this, NO_SUBSCRIBER_NO_RESULT, NO_REQUEST_NO_RESULT)) {
@@ -305,8 +271,8 @@ public class UnicastMonoProcessor<O> extends Mono<O>
         state = NO_REQUEST_HAS_RESULT;
       }
 
-
-      // check if state is with a result then there is a chance of immediate termination if there is no value
+      // check if state is with a result then there is a chance of immediate termination if there is
+      // no value
       // e.g. [onError / onComplete / dispose] only
       if (state == NO_REQUEST_HAS_RESULT && this.value == null) {
         this.actual = null;
@@ -323,72 +289,25 @@ public class UnicastMonoProcessor<O> extends Mono<O>
 
       // call onSubscribe if has value in the result or no result delivered so far
       actual.onSubscribe(this);
-
-      // update state to ensure we observe what happened after onSubscribe appeared
-      state = this.state;
-      // if FUSED_HAS_RESULT appeared then we have to notify Subscriber that there is a result which MUST be consumed
-      if (state == FUSED_HAS_RESULT) {
-        this.actual = null;
-        Throwable e = this.error;
-        if (e != null) {
-          // racing between cancel / dispose / complete
-          if (STATE.compareAndSet(this, FUSED_HAS_RESULT, FUSED_CONSUMED)) {
-            actual.onError(e);
-          }
-        } else {
-          O v = this.value;
-
-          if (v == null) {
-            if (STATE.compareAndSet(this, FUSED_HAS_RESULT, FUSED_CONSUMED)) {
-              actual.onComplete();
-            }
-          } else {
-            if (STATE.compareAndSet(this, FUSED_HAS_RESULT, FUSED_READY)) {
-              actual.onNext(v);
-              actual.onComplete();
-            }
-          }
-        }
-      }
     } else {
       Operators.error(
-              actual,
-              new IllegalStateException("UnicastMonoProcessor allows only a single Subscriber"));
+          actual,
+          new IllegalStateException("UnicastMonoProcessor allows only a single Subscriber"));
     }
-  }
-
-  @Override
-  public int requestFusion(int mode) {
-    if ((mode & ASYNC) != 0) {
-      int state = this.state;
-
-      // CAS LOOP since there is a racing with [onNext / onComplete / onError / dispose] which can appear at any moment
-      for (;;) {
-        if (state == NO_REQUEST_NO_RESULT && STATE.compareAndSet(this, NO_REQUEST_NO_RESULT, FUSED_EMPTY)) {
-          break;
-        }
-        if (state == NO_REQUEST_HAS_RESULT && STATE.compareAndSet(this, NO_REQUEST_HAS_RESULT, FUSED_HAS_RESULT)) {
-          break;
-        }
-
-        state = this.state;
-      }
-      return ASYNC;
-    }
-    return NONE;
   }
 
   @Override
   public final void request(long n) {
     if (Operators.validate(n)) {
-      for (;;) {
+      for (; ; ) {
         int s = state;
         // if the any bits 1-31 are set, we are either in fusion mode (FUSED_*)
         // or request has been called (HAS_REQUEST_*)
         if ((s & ~NO_REQUEST_HAS_RESULT) != 0) {
           return;
         }
-        if (s == NO_REQUEST_HAS_RESULT && STATE.compareAndSet(this, NO_REQUEST_HAS_RESULT, HAS_REQUEST_HAS_RESULT)) {
+        if (s == NO_REQUEST_HAS_RESULT
+            && STATE.compareAndSet(this, NO_REQUEST_HAS_RESULT, HAS_REQUEST_HAS_RESULT)) {
           O v = value;
           if (v != null) {
             value = null;
@@ -434,17 +353,6 @@ public class UnicastMonoProcessor<O> extends Mono<O>
     complete(new CancellationException("Disposed"));
   }
 
-  @Override
-  public O poll() {
-    if (STATE.compareAndSet(this, FUSED_READY, FUSED_CONSUMED)) {
-      O v = value;
-      value = null;
-      actual = null;
-      return v;
-    }
-    return null;
-  }
-
   /**
    * Returns the value that completed this {@link UnicastMonoProcessor}. Returns {@code null} if the
    * {@link UnicastMonoProcessor} has not been completed. If the {@link UnicastMonoProcessor} is
@@ -456,7 +364,7 @@ public class UnicastMonoProcessor<O> extends Mono<O>
    */
   @Nullable
   public O peek() {
-    if (!isDisposed()) {
+    if (isCancelled()) {
       return null;
     }
 
@@ -471,29 +379,6 @@ public class UnicastMonoProcessor<O> extends Mono<O>
     }
 
     return null;
-  }
-
-  @Override
-  public int size() {
-    int state = this.state;
-
-    if (state < FUSED_EMPTY && state % 2 == 1 && this.value != null) {
-      return 1;
-    } else if (this.value != null) {
-      return 1;
-    }
-
-    return 0;
-  }
-
-  @Override
-  public boolean isEmpty() {
-    return size() == 0;
-  }
-
-  @Override
-  public void clear() {
-    this.value = null;
   }
 
   /**
@@ -527,16 +412,6 @@ public class UnicastMonoProcessor<O> extends Mono<O>
   }
 
   /**
-   * Indicates whether this {@code UnicastMonoProcessor} has been interrupted via cancellation.
-   *
-   * @return {@code true} if this {@code UnicastMonoProcessor} is cancelled, {@code false}
-   *     otherwise.
-   */
-  public boolean isCancelled() {
-    return state == CANCELLED;
-  }
-
-  /**
    * Indicates whether this {@code UnicastMonoProcessor} has been completed with an error.
    *
    * @return {@code true} if this {@code UnicastMonoProcessor} was completed with an error, {@code
@@ -546,10 +421,25 @@ public class UnicastMonoProcessor<O> extends Mono<O>
     return getError() != null;
   }
 
+  /**
+   * Indicates whether this {@code UnicastMonoProcessor} has been interrupted via cancellation.
+   *
+   * @return {@code true} if this {@code UnicastMonoProcessor} is cancelled, {@code false}
+   *     otherwise.
+   */
+  public boolean isCancelled() {
+    return state == CANCELLED;
+  }
+
+  public final boolean isTerminated() {
+    int state = this.state;
+    return (state < CANCELLED && state % 2 == 1);
+  }
+
   @Override
   public boolean isDisposed() {
     int state = this.state;
-    return state == HAS_REQUEST_HAS_RESULT || state == CANCELLED || state == FUSED_CONSUMED;
+    return state == CANCELLED || (state < CANCELLED && state % 2 == 1);
   }
 
   @Override
@@ -559,7 +449,7 @@ public class UnicastMonoProcessor<O> extends Mono<O>
     int state = this.state;
 
     if (key == Attr.TERMINATED) {
-      return state == HAS_REQUEST_HAS_RESULT || state == FUSED_CONSUMED;
+      return (state < CANCELLED && state % 2 == 1);
     }
     if (key == Attr.PARENT) {
       return subscription;